Refactor authentication and user data handling
Replaced LoginHandler and UserData with a new Authenticator class that manages user authentication and stores user data. Updated Main and TransactionsHandler to use Authenticator for account validation and data access. Modified transaction methods to accept account index and user data as parameters, and added a balance display method to UI.

Also finalized the Transaction method of Balance inquiry, making it the first method in the TransactionHandler class to be finished.

That's all