Finished coding the logic of the whole program
I finally finished programming the logic of the whole program in accordance to the algorithm I wrote last few weeks ago hahaha.

All there is to it is to debug possible logical errors, as well as clean up the code to be more  readable for the human eye.

That's all.