finished coding the entire program
Finished coding the program, this as well as debug the program with the logical errors that I've encountered.

Final part of coding phase is:
COMMENTING AND DOCUMENTING USING JAVADOC

I'll be doing some documentation work tomorrow possibly for readability as well as to practice my documentation skills hahaha.

That's all.