Update Main.java
Corrected some logical errors in terms of returning the wrong values and missing the return command in some. I fixed **postTransactionAction** to where if the user chooses 1, corresponding to another transaction, and then logging out in the transaction menu, the user will now log out instead of accidentally ending the program.